Reintegrated session management
